Partially fixed issue 19 (backend fix complete, client-side fix in progress...)

+++ Sources/ProcessAJAX.php
@@ -259,8 +259,12 @@ if(isset($_POST['editprofilesubmit'])) {
 if(isset($_POST['deleteusersubmit'])) {
 	// Can the user do this?
 	if(permissions(2)) {
-		# Execute query to delete user
-		$dbh->query("DELETE FROM users WHERE id=".(int)$_POST['id']) or die(sqlite_error_string($dbh->lastError()));
+		# Check the user level of the user being deleted
+		$query = $dbh->query("SELECT role FROM users WHERE id=".(int)$_POST['id']) or die(sqlite_error_string($dbh->lastError()));
+		if(userFetch('role', 2) >= $query->fetchSingle()) {		
+			# Execute query to delete user
+			$dbh->query("DELETE FROM users WHERE id=".(int)$_POST['id']) or die(sqlite_error_string($dbh->lastError()));
+		}
 	}
 }
